## Release Notes: v2.8.0  (version bump)
This release is the first release in months that needed an update based on the changes in the enviroment. In general, the Litewallet code base is bloated with 1000's of lines of unused code and this makes debugging and adding new features very difficult. In addition, the requirements of iOS 14 and the plan to move from UIKit to a SwiftUI+Combine architecture means subsequent PRs will reflect a more streamlined codebase.

## Tech Debt
- Also have watch xcscheme added per upgrade
- Added Issue templates (#122)
	- Bug Report
	- Feature Request
- Updated key for RemoteRunnable in xcscheme
- remove unused fonts
- run code on main thread
- update cocoapods firebase
- update addresses + event

## Improvements
* compute QR code only once
	- Soft launch for Import QR Code

## Bugfixes
* fix crash: window required

## Temporary Workarounds
The proxy server the supports buying LTC is being refactored and so no mobile clients can use the Simplex service
	* Disabled Simplex / Added temp message
		- Was causing crashes when return messages were happening.
		- Will update in the next cycle with an updated UI
		- Added a Analytics:  DID TAP BUY TAB marker
		- Added Coming Soon label to the tableview background view for a temporary sign
		- Added all translation for the coming soon message

## Authors
Kerry Washington <mrkerrywashington@icloud.com>
Mohamed Barry <mosadialiou@gmail.com>